+++ src/App.css
@@ -1,38 +1,31 @@
-.App {
-  text-align: center;
-}
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
+.global {
+    text-align: left;
+    position: relative;
+    background: radial-gradient(rgba(0, 0, 0, 0.5) 40%, rgba(0, 0, 0, 0.3))no-repeat,
+                url('images/globe.jpg') no-repeat;
+    color: #ffffff;
+    background-size: cover;
+    background-color: #464646;
+    background-position: center;
+    background-attachment: fixed;
+  } 
 
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
+  .global h1 {
+      margin: 0 0 0.5em 0 ;
   }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
+  .global p {
+      margin: 0;
   }
-  to {
-    transform: rotate(360deg);
+
+  .global .global-container, .global h1 {
+      width: 90%;
+      margin: 0 auto;
   }
-}
+
+  .global svg {
+      position: absolute;
+      right: 2em;
+      font-size: 1.5em;
+      bottom: 0.5em;
+  }
