+++ src/App.js
@@ -1,14 +1,28 @@
-import React from 'react';
+import React, { useEffect, Suspense, Fragment } from 'react';
+import { connect } from 'react-redux';
 import './App.css';
 
 import Home from './components/Home';
+import * as ActionCreators from './actions';
 
-function App() {
+function App(props) {
+  useEffect(() => { 
+    props.fetchData();
+  }, []);
+
   return (
-    <div className="App">      
-      <Home/>
-    </div>
+    <Fragment>
+      <Suspense fallback={<div>Loading...</div>}>
+        <div className="App">      
+          <Home/>
+        </div>
+      </Suspense>
+    </Fragment>
   );
 }
 
-export default App;
+const mapStateToProps = (state => {
+  return state;
+});
+
+export default connect(mapStateToProps, ActionCreators)(App);
